+++ app/src/main/java/camp/nextstep/service/TxUserService.java
@@ -1,0 +1,44 @@
+package camp.nextstep.service;
+
+import camp.nextstep.domain.User;
+import camp.nextstep.jdbc.datasource.TransactionManager;
+
+public class TxUserService implements UserService {
+
+    private final UserService userService;
+    private final TransactionManager transactionManager;
+
+    public TxUserService(UserService userService, TransactionManager transactionManager) {
+        this.userService = userService;
+        this.transactionManager = transactionManager;
+    }
+
+    @Override
+    public User findByAccount(String account) {
+        return userService.findByAccount(account);
+    }
+
+    @Override
+    public User findById(long id) {
+        return userService.findById(id);
+    }
+
+    @Override
+    public void save(User user) {
+        userService.save(user);
+    }
+
+    @Override
+    public void changePassword(long id, String newPassword, String createdBy) {
+        transactionManager.begin();
+        try {
+            userService.changePassword(id, newPassword, createdBy);
+            transactionManager.commit();
+        } catch (Exception e) {
+            transactionManager.rollback();
+            throw e;
+        } finally {
+            transactionManager.close();
+        }
+    }
+}
