+++ app/src/test/java/camp/nextstep/service/UserServiceTest.java
@@ -9,13 +9,11 @@ import camp.nextstep.jdbc.core.JdbcTemplate;
 import camp.nextstep.jdbc.datasource.TransactionManager;
 import camp.nextstep.support.jdbc.init.DatabasePopulatorUtils;
 import java.sql.SQLException;
-import javax.sql.DataSource;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 
 @Disabled
 class UserServiceTest {
@@ -40,7 +38,7 @@ class UserServiceTest {
     @Test
     void testChangePassword() throws SQLException {
         final var userHistoryDao = new UserHistoryDao(jdbcTemplate);
-        final var userService = new UserService(userDao, userHistoryDao, transactionManager);
+        final var userService = new TxUserService(new AppUserService(userDao, userHistoryDao), transactionManager);
 
         final var newPassword = "qqqqq";
         final var createBy = "gugu";
@@ -55,7 +53,7 @@ class UserServiceTest {
     void testTransactionRollback() {
         // 트랜잭션 롤백 테스트를 위해 mock으로 교체
         final var userHistoryDao = new MockUserHistoryDao(jdbcTemplate);
-        final var userService = new UserService(userDao, userHistoryDao, transactionManager);
+        final var userService = new TxUserService(new AppUserService(userDao, userHistoryDao), transactionManager);
 
         final var newPassword = "newPassword";
         final var createBy = "gugu";
